Put CanMessage/CanSignal inherited classes in canmessage. Delete travis yml and create mkdocs
