+++ mkdocs.yml
@@ -1,0 +1,10 @@
+site_name: caroa04
+site_description: Library to control the CAROA04 CAN-IO expander device from eletechsup.
+site_author: Rodolphe Mete Soyding
+
+theme: readthedocs
+
+repo_url: https://github.com/supermete/caroa04
+
+pages:
+- Home: index.md
